Widen the Inkling sconv delta path to FP32 on Volta
The NaN was an FP16 overflow, one step earlier than the attention kernel I
had been suspecting. Per-rank probes at layer 10 put the first non-finite
value on TP rank 4:

  rank 4:  after_rs 7748  -> after_sconv inf
  ranks 0-3,5-7: after_rs 471-1047 -> after_sconv 711-3612 (finite)
                 -> after_ag inf (all-gather spreads rank 4's inf)
                 -> rmsnorm(inf) -> attn_in NaN -> qkvr NaN

An all-gather only concatenates, so it cannot manufacture an inf; rank 4
owns the hidden-dim slice holding an outlier channel and its shard is 7-16x
every other rank's. The captured dump agrees: q is NaN only at token 0 and
KV page 17 is NaN in slots 0-3, i.e. NaN arriving as input rather than
being produced by attention.

The conv kernel already accumulates in FP32 (acc = tl.zeros(..., tl.float32),
all loads .to(tl.float32)); the single lossy step is the final store
acc.to(out_ptr.dtype.element_ty). So the fp32 result genuinely exceeds
65504 and the fix is to widen the *output*, not the accumulator. Inkling is
BF16-native (max ~3.4e38) and Volta has no BF16.

fused_sconv takes out_dtype, the fallback RS/sconv/AG path runs the conv and
the gather in FP32, and add_rmsnorm's y now follows the norm weight rather
than the delta so sublayers and the quantized GEMMs still see FP16.

This also corrects the add_rmsnorm comment, which credited an earlier
residual-overflow theory I retracted: the residual was never the tensor that
overflowed, which is why making it FP32 changed nothing.

+++ vllm/models/inkling/nvidia/model.py
@@ -136,7 +136,9 @@ def _sconv_add_norm(
     _p("delta_summed", delta)
     shard = tensor_model_parallel_reduce_scatter(delta, dim=-1)
     _p("after_rs", shard)
-    shard = sconv(shard.contiguous(), positions)
+    # FP32 out: the conv accumulates in FP32 already, but narrowing the store
+    # to FP16 overflows on this checkpoint's outlier channels (see add_rmsnorm).
+    shard = sconv(shard.contiguous(), positions, out_dtype=torch.float32)
     _p("after_sconv", shard)
     full = tensor_model_parallel_all_gather(shard, dim=-1)
     _p("after_ag", full)
